+++ sdk/messaging/azservicebus/admin/admin_client_test.go
@@ -35,7 +35,7 @@ func TestAdminClient_UsingIdentity(t *testing.T) {
 	require.NoError(t, err)
 
 	queueName := fmt.Sprintf("queue-%X", time.Now().UnixNano())
-	props, err := adminClient.CreateQueue(context.Background(), queueName, nil, nil)
+	props, err := adminClient.CreateQueue(context.Background(), queueName, nil)
 	require.NoError(t, err)
 	require.EqualValues(t, 10, *props.MaxDeliveryCount)
 
@@ -77,23 +77,25 @@ func TestAdminClient_QueueWithMaxValues(t *testing.T) {
 
 	queueName := fmt.Sprintf("queue-%X", time.Now().UnixNano())
 
-	_, err = adminClient.CreateQueue(context.Background(), queueName, &QueueProperties{
-		LockDuration: to.Ptr("PT45S"),
-		// when you enable partitioning Service Bus will automatically create 16 partitions, each with the size
-		// of MaxSizeInMegabytes. This means when we retrieve this queue we'll get 16*4096 as the size (ie: 64GB)
-		EnablePartitioning:                  to.Ptr(true),
-		MaxSizeInMegabytes:                  to.Ptr(int32(4096)),
-		RequiresDuplicateDetection:          to.Ptr(true),
-		RequiresSession:                     to.Ptr(true),
-		DefaultMessageTimeToLive:            MaxTimeSpanForTests,
-		DeadLetteringOnMessageExpiration:    to.Ptr(true),
-		DuplicateDetectionHistoryTimeWindow: to.Ptr("PT4H"),
-		MaxDeliveryCount:                    to.Ptr(int32(100)),
-		EnableBatchedOperations:             to.Ptr(false),
-		Status:                              &es,
-		AutoDeleteOnIdle:                    MaxTimeSpanForTests,
-		UserMetadata:                        to.Ptr("some metadata"),
-	}, nil)
+	_, err = adminClient.CreateQueue(context.Background(), queueName, &CreateQueueOptions{
+		Properties: &QueueProperties{
+			LockDuration: to.Ptr("PT45S"),
+			// when you enable partitioning Service Bus will automatically create 16 partitions, each with the size
+			// of MaxSizeInMegabytes. This means when we retrieve this queue we'll get 16*4096 as the size (ie: 64GB)
+			EnablePartitioning:                  to.Ptr(true),
+			MaxSizeInMegabytes:                  to.Ptr(int32(4096)),
+			RequiresDuplicateDetection:          to.Ptr(true),
+			RequiresSession:                     to.Ptr(true),
+			DefaultMessageTimeToLive:            MaxTimeSpanForTests,
+			DeadLetteringOnMessageExpiration:    to.Ptr(true),
+			DuplicateDetectionHistoryTimeWindow: to.Ptr("PT4H"),
+			MaxDeliveryCount:                    to.Ptr(int32(100)),
+			EnableBatchedOperations:             to.Ptr(false),
+			Status:                              &es,
+			AutoDeleteOnIdle:                    MaxTimeSpanForTests,
+			UserMetadata:                        to.Ptr("some metadata"),
+		},
+	})
 	require.NoError(t, err)
 
 	defer deleteQueue(t, adminClient, queueName)
@@ -138,22 +140,24 @@ func TestAdminClient_CreateQueue(t *testing.T) {
 	queueName := fmt.Sprintf("queue-%X", time.Now().UnixNano())
 
 	es := EntityStatusReceiveDisabled
-	createResp, err := adminClient.CreateQueue(context.Background(), queueName, &QueueProperties{
-		LockDuration: to.Ptr("PT45S"),
-		// when you enable partitioning Service Bus will automatically create 16 partitions, each with the size
-		// of MaxSizeInMegabytes. This means when we retrieve this queue we'll get 16*4096 as the size (ie: 64GB)
-		EnablePartitioning:                  to.Ptr(true),
-		MaxSizeInMegabytes:                  to.Ptr(int32(4096)),
-		RequiresDuplicateDetection:          to.Ptr(true),
-		RequiresSession:                     to.Ptr(true),
-		DefaultMessageTimeToLive:            to.Ptr("PT6H"),
-		DeadLetteringOnMessageExpiration:    to.Ptr(true),
-		DuplicateDetectionHistoryTimeWindow: to.Ptr("PT4H"),
-		MaxDeliveryCount:                    to.Ptr(int32(100)),
-		EnableBatchedOperations:             to.Ptr(false),
-		Status:                              &es,
-		AutoDeleteOnIdle:                    to.Ptr("PT10M"),
-	}, nil)
+	createResp, err := adminClient.CreateQueue(context.Background(), queueName, &CreateQueueOptions{
+		Properties: &QueueProperties{
+			LockDuration: to.Ptr("PT45S"),
+			// when you enable partitioning Service Bus will automatically create 16 partitions, each with the size
+			// of MaxSizeInMegabytes. This means when we retrieve this queue we'll get 16*4096 as the size (ie: 64GB)
+			EnablePartitioning:                  to.Ptr(true),
+			MaxSizeInMegabytes:                  to.Ptr(int32(4096)),
+			RequiresDuplicateDetection:          to.Ptr(true),
+			RequiresSession:                     to.Ptr(true),
+			DefaultMessageTimeToLive:            to.Ptr("PT6H"),
+			DeadLetteringOnMessageExpiration:    to.Ptr(true),
+			DuplicateDetectionHistoryTimeWindow: to.Ptr("PT4H"),
+			MaxDeliveryCount:                    to.Ptr(int32(100)),
+			EnableBatchedOperations:             to.Ptr(false),
+			Status:                              &es,
+			AutoDeleteOnIdle:                    to.Ptr("PT10M"),
+		},
+	})
 	require.NoError(t, err)
 
 	defer func() {
@@ -187,7 +191,7 @@ func TestAdminClient_UpdateQueue(t *testing.T) {
 	require.NoError(t, err)
 
 	queueName := fmt.Sprintf("queue-%X", time.Now().UnixNano())
-	createdProps, err := adminClient.CreateQueue(context.Background(), queueName, nil, nil)
+	createdProps, err := adminClient.CreateQueue(context.Background(), queueName, nil)
 	require.NoError(t, err)
 
 	defer func() {
@@ -225,9 +229,11 @@ func TestAdminClient_ListQueues(t *testing.T) {
 		queueName := strings.ToLower(fmt.Sprintf("list-queues-%d-%X", i, now))
 		expectedQueues = append(expectedQueues, queueName)
 
-		_, err = adminClient.CreateQueue(context.Background(), queueName, &QueueProperties{
-			MaxDeliveryCount: to.Ptr(int32(int32(i + 10))),
-		}, nil)
+		_, err = adminClient.CreateQueue(context.Background(), queueName, &CreateQueueOptions{
+			Properties: &QueueProperties{
+				MaxDeliveryCount: to.Ptr(int32(int32(i + 10))),
+			},
+		})
 		require.NoError(t, err)
 
 		defer deleteQueue(t, adminClient, queueName)
@@ -241,7 +247,7 @@ func TestAdminClient_ListQueues(t *testing.T) {
 		page, err := pager.NextPage(context.Background())
 		require.NoError(t, err)
 
-		for _, props := range page.Items {
+		for _, props := range page.Queues {
 			all[props.QueueName] = props
 		}
 	}
@@ -266,9 +272,10 @@ func TestAdminClient_ListQueuesRuntimeProperties(t *testing.T) {
 		queueName := strings.ToLower(fmt.Sprintf("list-queuert-%d-%X", i, now))
 		expectedQueues = append(expectedQueues, queueName)
 
-		_, err = adminClient.CreateQueue(context.Background(), queueName, &QueueProperties{
-			MaxDeliveryCount: to.Ptr(int32(int32(i + 10))),
-		}, nil)
+		_, err = adminClient.CreateQueue(context.Background(), queueName, &CreateQueueOptions{
+			Properties: &QueueProperties{
+				MaxDeliveryCount: to.Ptr(int32(int32(i + 10))),
+			}})
 		require.NoError(t, err)
 
 		defer deleteQueue(t, adminClient, queueName)
@@ -288,9 +295,9 @@ func TestAdminClient_ListQueuesRuntimeProperties(t *testing.T) {
 		resp, err := pager.NextPage(context.Background())
 		require.NoError(t, err)
 
-		require.LessOrEqual(t, len(resp.Items), 2)
+		require.LessOrEqual(t, len(resp.QueueRuntimeProperties), 2)
 
-		for _, queueRuntimeItem := range resp.Items {
+		for _, queueRuntimeItem := range resp.QueueRuntimeProperties {
 			// _, exists := all[queueRuntimeItem.QueueName]
 			// require.False(t, exists, fmt.Sprintf("Each queue result should be unique but found more than one of '%s'", queueRuntimeItem.QueueName))
 			all[queueRuntimeItem.QueueName] = queueRuntimeItem
@@ -360,7 +367,7 @@ func TestAdminClient_TopicAndSubscription(t *testing.T) {
 	subscriptionName := fmt.Sprintf("sub-%X", time.Now().UnixNano())
 	forwardToQueueName := fmt.Sprintf("queue-fwd-%X", time.Now().UnixNano())
 
-	_, err = adminClient.CreateQueue(context.Background(), forwardToQueueName, nil, nil)
+	_, err = adminClient.CreateQueue(context.Background(), forwardToQueueName, nil)
 	require.NoError(t, err)
 
 	defer deleteQueue(t, adminClient, forwardToQueueName)
@@ -368,18 +375,19 @@ func TestAdminClient_TopicAndSubscription(t *testing.T) {
 	status := EntityStatusActive
 
 	// check topic properties, existence
-	addResp, err := adminClient.CreateTopic(context.Background(), topicName, &TopicProperties{
-		EnablePartitioning:                  to.Ptr(true),
-		MaxSizeInMegabytes:                  to.Ptr(int32(2048)),
-		RequiresDuplicateDetection:          to.Ptr(true),
-		DefaultMessageTimeToLive:            to.Ptr("PT3M"),
-		DuplicateDetectionHistoryTimeWindow: to.Ptr("PT4M"),
-		EnableBatchedOperations:             to.Ptr(true),
-		Status:                              &status,
-		AutoDeleteOnIdle:                    to.Ptr("PT7M"),
-		SupportOrdering:                     to.Ptr(true),
-		UserMetadata:                        to.Ptr("user metadata"),
-	}, nil)
+	addResp, err := adminClient.CreateTopic(context.Background(), topicName, &CreateTopicOptions{
+		Properties: &TopicProperties{
+			EnablePartitioning:                  to.Ptr(true),
+			MaxSizeInMegabytes:                  to.Ptr(int32(2048)),
+			RequiresDuplicateDetection:          to.Ptr(true),
+			DefaultMessageTimeToLive:            to.Ptr("PT3M"),
+			DuplicateDetectionHistoryTimeWindow: to.Ptr("PT4M"),
+			EnableBatchedOperations:             to.Ptr(true),
+			Status:                              &status,
+			AutoDeleteOnIdle:                    to.Ptr("PT7M"),
+			SupportOrdering:                     to.Ptr(true),
+			UserMetadata:                        to.Ptr("user metadata"),
+		}})
 	require.NoError(t, err)
 
 	defer deleteTopic(t, adminClient, topicName)
@@ -423,20 +431,22 @@ func TestAdminClient_TopicAndSubscription(t *testing.T) {
 	require.Zero(t, runtimeResp.ScheduledMessageCount)
 	require.Zero(t, runtimeResp.SizeInBytes)
 
-	addSubWithPropsResp, err := adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, &SubscriptionProperties{
-		LockDuration:                                    to.Ptr("PT3M"),
-		RequiresSession:                                 to.Ptr(false),
-		DefaultMessageTimeToLive:                        to.Ptr("PT7M"),
-		DeadLetteringOnMessageExpiration:                to.Ptr(true),
-		EnableDeadLetteringOnFilterEvaluationExceptions: to.Ptr(false),
-		MaxDeliveryCount:                                to.Ptr(int32(11)),
-		Status:                                          &status,
-		// ForwardTo:                     &forwardToQueueName,
-		// ForwardDeadLetteredMessagesTo: &forwardToQueueName,
-		EnableBatchedOperations: to.Ptr(false),
-		AutoDeleteOnIdle:        to.Ptr("PT11M"),
-		UserMetadata:            to.Ptr("user metadata"),
-	}, nil)
+	addSubWithPropsResp, err := adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, &CreateSubscriptionOptions{
+		Properties: &SubscriptionProperties{
+			LockDuration:                                    to.Ptr("PT3M"),
+			RequiresSession:                                 to.Ptr(false),
+			DefaultMessageTimeToLive:                        to.Ptr("PT7M"),
+			DeadLetteringOnMessageExpiration:                to.Ptr(true),
+			EnableDeadLetteringOnFilterEvaluationExceptions: to.Ptr(false),
+			MaxDeliveryCount:                                to.Ptr(int32(11)),
+			Status:                                          &status,
+			// ForwardTo:                     &forwardToQueueName,
+			// ForwardDeadLetteredMessagesTo: &forwardToQueueName,
+			EnableBatchedOperations: to.Ptr(false),
+			AutoDeleteOnIdle:        to.Ptr("PT11M"),
+			UserMetadata:            to.Ptr("user metadata"),
+		},
+	})
 	require.NoError(t, err)
 
 	defer deleteSubscription(t, adminClient, topicName, subscriptionName)
@@ -462,7 +472,7 @@ func TestAdminClient_UpdateTopic(t *testing.T) {
 	require.NoError(t, err)
 
 	topicName := fmt.Sprintf("topic-%X", time.Now().UnixNano())
-	addResp, err := adminClient.CreateTopic(context.Background(), topicName, nil, nil)
+	addResp, err := adminClient.CreateTopic(context.Background(), topicName, nil)
 	require.NoError(t, err)
 
 	defer deleteTopic(t, adminClient, topicName)
@@ -503,9 +513,11 @@ func TestAdminClient_ListTopics(t *testing.T) {
 
 		go func(i int) {
 			defer wg.Done()
-			_, err := adminClient.CreateTopic(context.Background(), topicName, &TopicProperties{
-				DefaultMessageTimeToLive: to.Ptr(fmt.Sprintf("PT%dM", i+1)),
-			}, nil)
+			_, err := adminClient.CreateTopic(context.Background(), topicName, &CreateTopicOptions{
+				Properties: &TopicProperties{
+					DefaultMessageTimeToLive: to.Ptr(fmt.Sprintf("PT%dM", i+1)),
+				},
+			})
 			require.NoError(t, err)
 		}(i)
 
@@ -528,9 +540,9 @@ func TestAdminClient_ListTopics(t *testing.T) {
 
 		times++
 
-		require.LessOrEqual(t, len(page.Items), 2)
+		require.LessOrEqual(t, len(page.Topics), 2)
 
-		for _, topicItem := range page.Items {
+		for _, topicItem := range page.Topics {
 			// _, exists := all[topicItem.TopicName]
 			// require.False(t, exists, fmt.Sprintf("Each topic result should be unique but found more than one of '%s'", topicItem.TopicName))
 			all[topicItem.TopicName] = topicItem
@@ -559,7 +571,7 @@ func TestAdminClient_ListTopicsRuntimeProperties(t *testing.T) {
 		topicName := strings.ToLower(fmt.Sprintf("list-topicrt-%d-%X", i, now))
 		expectedTopics = append(expectedTopics, topicName)
 
-		_, err = adminClient.CreateTopic(context.Background(), topicName, nil, nil)
+		_, err = adminClient.CreateTopic(context.Background(), topicName, nil)
 		require.NoError(t, err)
 
 		defer deleteTopic(t, adminClient, topicName)
@@ -579,9 +591,9 @@ func TestAdminClient_ListTopicsRuntimeProperties(t *testing.T) {
 
 		times++
 
-		require.LessOrEqual(t, len(page.Items), 2)
+		require.LessOrEqual(t, len(page.TopicRuntimeProperties), 2)
 
-		for _, item := range page.Items {
+		for _, item := range page.TopicRuntimeProperties {
 			// _, exists := all[item.TopicName]
 			// require.False(t, exists, fmt.Sprintf("Each topic result should be unique but found more than one of '%s'", item.TopicName))
 			all[item.TopicName] = item
@@ -606,7 +618,7 @@ func TestAdminClient_ListSubscriptions(t *testing.T) {
 	now := time.Now().UnixNano()
 	topicName := strings.ToLower(fmt.Sprintf("listsub-%X", now))
 
-	_, err = adminClient.CreateTopic(context.Background(), topicName, nil, nil)
+	_, err = adminClient.CreateTopic(context.Background(), topicName, nil)
 	require.NoError(t, err)
 
 	defer deleteTopic(t, adminClient, topicName)
@@ -617,9 +629,11 @@ func TestAdminClient_ListSubscriptions(t *testing.T) {
 		subName := strings.ToLower(fmt.Sprintf("sub-%d-%X", i, now))
 		expectedSubscriptions = append(expectedSubscriptions, subName)
 
-		_, err = adminClient.CreateSubscription(context.Background(), topicName, subName, &SubscriptionProperties{
-			DefaultMessageTimeToLive: to.Ptr(fmt.Sprintf("PT%dM", i+1)),
-		}, nil)
+		_, err = adminClient.CreateSubscription(context.Background(), topicName, subName, &CreateSubscriptionOptions{
+			Properties: &SubscriptionProperties{
+				DefaultMessageTimeToLive: to.Ptr(fmt.Sprintf("PT%dM", i+1)),
+			},
+		})
 		require.NoError(t, err)
 
 		defer deleteSubscription(t, adminClient, topicName, subName)
@@ -638,9 +652,9 @@ func TestAdminClient_ListSubscriptions(t *testing.T) {
 		page, err := pager.NextPage(context.Background())
 		require.NoError(t, err)
 
-		require.LessOrEqual(t, len(page.Items), 2)
+		require.LessOrEqual(t, len(page.Subscriptions), 2)
 
-		for _, item := range page.Items {
+		for _, item := range page.Subscriptions {
 			// _, exists := all[item.SubscriptionName]
 			// require.False(t, exists, fmt.Sprintf("Each subscription result should be unique but found more than one of '%s'", item.SubscriptionName))
 			all[item.SubscriptionName] = item
@@ -666,7 +680,7 @@ func TestAdminClient_ListSubscriptionRuntimeProperties(t *testing.T) {
 	now := time.Now().UnixNano()
 	topicName := strings.ToLower(fmt.Sprintf("listsubrt-%X", now))
 
-	_, err = adminClient.CreateTopic(context.Background(), topicName, nil, nil)
+	_, err = adminClient.CreateTopic(context.Background(), topicName, nil)
 	require.NoError(t, err)
 
 	var expectedSubs []string
@@ -675,7 +689,7 @@ func TestAdminClient_ListSubscriptionRuntimeProperties(t *testing.T) {
 		subscriptionName := strings.ToLower(fmt.Sprintf("sub-%d-%X", i, now))
 		expectedSubs = append(expectedSubs, subscriptionName)
 
-		_, err = adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, nil, nil)
+		_, err = adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, nil)
 		require.NoError(t, err)
 
 		defer deleteSubscription(t, adminClient, topicName, subscriptionName)
@@ -693,9 +707,9 @@ func TestAdminClient_ListSubscriptionRuntimeProperties(t *testing.T) {
 		page, err := pager.NextPage(context.Background())
 		require.NoError(t, err)
 
-		require.LessOrEqual(t, len(page.Items), 2)
+		require.LessOrEqual(t, len(page.SubscriptionRuntimeProperties), 2)
 
-		for _, subItem := range page.Items {
+		for _, subItem := range page.SubscriptionRuntimeProperties {
 			// _, exists := all[subItem.SubscriptionName]
 			// require.False(t, exists, fmt.Sprintf("Each subscription result should be unique but found more than one of '%s'", subItem.SubscriptionName))
 			all[subItem.SubscriptionName] = subItem
@@ -725,13 +739,13 @@ func TestAdminClient_UpdateSubscription(t *testing.T) {
 	require.NoError(t, err)
 
 	topicName := fmt.Sprintf("topic-%X", time.Now().UnixNano())
-	_, err = adminClient.CreateTopic(context.Background(), topicName, nil, nil)
+	_, err = adminClient.CreateTopic(context.Background(), topicName, nil)
 	require.NoError(t, err)
 
 	defer deleteTopic(t, adminClient, topicName)
 
 	subscriptionName := fmt.Sprintf("sub-%X", time.Now().UnixNano())
-	addResp, err := adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, nil, nil)
+	addResp, err := adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, nil)
 	require.NoError(t, err)
 
 	defer deleteSubscription(t, adminClient, topicName, subscriptionName)
@@ -778,12 +792,12 @@ func TestAdminClient_LackPermissions_Queue(t *testing.T) {
 
 	pager := testData.Client.ListQueues(nil)
 	page, err := pager.NextPage(context.Background())
-	require.Empty(t, page.Items)
+	require.Empty(t, page.Queues)
 	require.Contains(t, err.Error(), "Manage,EntityRead claims required for this operation")
 	require.ErrorAs(t, err, &re)
 	require.EqualValues(t, 401, re.StatusCode)
 
-	_, err = testData.Client.CreateQueue(ctx, "canneverbecreated", nil, nil)
+	_, err = testData.Client.CreateQueue(ctx, "canneverbecreated", nil)
 	require.Contains(t, err.Error(), "Authorization failed for specified action: Manage,EntityWrite")
 	require.ErrorAs(t, err, &re)
 	require.EqualValues(t, 401, re.StatusCode)
@@ -826,7 +840,7 @@ func TestAdminClient_LackPermissions_Topic(t *testing.T) {
 	require.ErrorAs(t, err, &asResponseErr)
 	require.EqualValues(t, 401, asResponseErr.StatusCode)
 
-	_, err = testData.Client.CreateTopic(ctx, "canneverbecreated", nil, nil)
+	_, err = testData.Client.CreateTopic(ctx, "canneverbecreated", nil)
 	require.Contains(t, err.Error(), "Authorization failed for specified action")
 	require.ErrorAs(t, err, &asResponseErr)
 	require.EqualValues(t, 401, asResponseErr.StatusCode)
@@ -875,7 +889,7 @@ func TestAdminClient_NotFound(t *testing.T) {
 	nanoSeconds := time.Now().UnixNano()
 	topicName := fmt.Sprintf("topic-%d", nanoSeconds)
 
-	_, err = adminClient.CreateTopic(context.Background(), topicName, nil, nil)
+	_, err = adminClient.CreateTopic(context.Background(), topicName, nil)
 	require.NoError(t, err)
 
 	defer deleteTopic(t, adminClient, topicName)
@@ -905,7 +919,7 @@ func TestAdminClient_LackPermissions_Subscription(t *testing.T) {
 	_, err = pager.NextPage(context.Background())
 	require.Contains(t, err.Error(), "401 SubCode=40100: Unauthorized : Unauthorized access for 'EnumerateSubscriptions' operation")
 
-	_, err = testData.Client.CreateSubscription(ctx, testData.TopicName, "canneverbecreated", nil, nil)
+	_, err = testData.Client.CreateSubscription(ctx, testData.TopicName, "canneverbecreated", nil)
 	require.Contains(t, err.Error(), "401 SubCode=40100: Unauthorized : Unauthorized access for 'CreateOrUpdateSubscription'")
 
 	_, err = testData.Client.UpdateSubscription(ctx, testData.TopicName, "canneverbecreated", SubscriptionProperties{}, nil)
@@ -1094,13 +1108,13 @@ func setupLowPrivTest(t *testing.T) *struct {
 	// create some entities that we need (there's a diff between something not being
 	// found and something failing because of lack of authorization)
 	cleanup := func() func() {
-		_, err = adminClient.CreateQueue(context.Background(), queueName, nil, nil)
+		_, err = adminClient.CreateQueue(context.Background(), queueName, nil)
 		require.NoError(t, err)
 
-		_, err = adminClient.CreateTopic(context.Background(), topicName, nil, nil)
+		_, err = adminClient.CreateTopic(context.Background(), topicName, nil)
 		require.NoError(t, err)
 
-		_, err = adminClient.CreateSubscription(context.Background(), topicName, subName, nil, nil)
+		_, err = adminClient.CreateSubscription(context.Background(), topicName, subName, nil)
 		require.NoError(t, err)
 
 		// _, err = sm.PutRule(context.Background(), subName, ruleName, TrueFilter{})